# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Production Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/tasks:
    get:
      operationId: listProjectTasks
      tags: [tasks]
      summary: List production tasks for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionTask"
...
  /tasks:
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a production task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionTask"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a production task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionTask"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a production task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ProductionTask:
      type: object
      required: [id, projectId, title, status, createdAt]
      properties:
        id:
          type: number
        projectId:
          type: number
        title:
          type: string
        status:
          type: string
        assignedActorId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [projectId, title]
      properties:
        projectId:
          type: number
        title:
          type: string
          minLength: 1
        status:
          type: string
        assignedActorId:
          type: number
        dueDate:
          type: string
    TaskPatch:
      type: object
      properties:
        title:
          type: string
        status:
          type: string
        assignedActorId:
          type: number
        dueDate:
          type: string
...